% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-token.R
\name{as_azure_token}
\alias{as_azure_token}
\title{Convert a token to the AzureAuth format}
\usage{
as_azure_token(token, scope = NULL)
}
\arguments{
\item{token}{An \link[httr2:oauth_token]{httr2::oauth_token}.}

\item{scope}{The Azure scope(s) this token is associated with, or \code{NULL} if
this is unknown.}
}
\value{
An R6 class that inherits from \code{AzureToken}.
}
\description{
Wraps an \link[httr2:oauth_token]{httr2::oauth_token} in an \link[AzureAuth:AzureToken]{AzureAuth::AzureToken} for use with the
AzureR package ecosystem.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
token <- NULL
if (has_viewer_token()) {
  token <- as_azure_token(connect_viewer_token())
}

# For example, when using the Microsoft365R package:
# Microsoft365R::list_sharepoint_sites(token = token)
\dontshow{\}) # examplesIf}
}
