% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcreds-package.R
\docType{package}
\name{connectcreds-package}
\alias{connectcreds}
\alias{connectcreds-package}
\title{connectcreds: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Aaron Jacobs \email{aaron.jacobs@posit.co}

}
\keyword{internal}
