% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcreds-package.R
\docType{package}
\name{connectcreds-package}
\alias{connectcreds}
\alias{connectcreds-package}
\title{connectcreds: Manage 'OAuth' Credentials from 'Posit Connect'}
\description{
A toolkit for making use of credentials mediated by 'Posit Connect'. It handles the details of communicating with Connect's API correctly, 'OAuth' token caching, and refresh behaviour.
}
\author{
\strong{Maintainer}: Aaron Jacobs \email{aaron.jacobs@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
